add more documentation

+++ screen.cpp
@@ -40,14 +40,14 @@ void Screen::draw_shape(const Shape& shape) {
     if (size.x / size.y > (double)width / height) {
         // we need to add black bars at the top and bottom
         // ekkor a fekete csíkokat alulra és felülre kell tenni
-        start_y = (height - width * size.y / size.x) / 2.0;
+        start_y = (unsigned)((height - width * size.y / size.x) / 2.0);
         max_height -= start_y;
     // if the actual size is wider than the simulated
     // ha az igazi képméret szélesebb, mint a szimulált
     } else {
         // we need to add black bars at the left and right
         // ekkor a fekete csíkokat jobbra és balra kell tenni
-        start_x = (width - height * size.x / size.y) / 2.0;
+        start_x = (unsigned)((width - height * size.x / size.y) / 2.0);
         max_width -= start_x;
     }
     // végigmegyünk azokon az igazi, képernyőn levő képpontokon, ahova tényleg rajzolni szeretnénk
@@ -83,7 +83,7 @@ void Screen::render(std::ostream& out) const {
         for (unsigned x = 0; x < width; ++x) {
             bool up = idx(x, 2 * y);
             bool down = idx(x, 2 * y + 1);
-            out << getblock((Block)(2 * up + down));
+            out << getblock((Block)(2 * (int)up + (int)down));
         }
         out << '\n';
     }
